Switched ModsContainer over to DB data
